Move markdown and dismissible alert (fix space problems)
