+++ client/web-sveltekit/src/lib/wildcard/Markdown.module.scss
@@ -1,3 +1,9 @@
+// Since all elements in Markdown components are dynamically
+// generated/provided from consumer all styles can't be set
+// statically with svelte style blocks. The easiness way to work
+// around this problem is CSS module instead of adding :global
+// to each selector (:global {} scoping doesn't work in svelte styles)
+
 .markdown {
     // marked wraps all content in <p>, even if there is only one child.
     // If that is the case, treat the <p> as if it wasn't there
