+++ client/web-sveltekit/src/lib/wildcard/index.ts
@@ -10,3 +10,4 @@ export { default as Submenu } from './menu/Submenu.svelte'
 export { default as MenuRadioGroup } from './menu/MenuRadioGroup.svelte'
 export { getFileIconInfo } from '@sourcegraph/wildcard/src/components/Icon'
 export { default as Alert } from './Alert.svelte'
+export {default as Markdown} from './Markdown.svelte'
