+++ client/web-sveltekit/vite.config.ts
@@ -21,7 +21,9 @@ export default defineConfig(({ mode }) => {
                       'process.platform': '"browser"',
                       'process.env.VITEST': 'null',
                       'process.env.NODE_ENV': `"${mode}"`,
-                    'process.env': `{ "SOURCEGRAPH_API_URL": "${process.env.SOURCEGRAPH_API_URL}" }`,
+                    'process.env': `{ "SOURCEGRAPH_API_URL": ${
+                        process.env.SOURCEGRAPH_API_URL ? '${process.env.SOURCEGRAPH_API_URL}' : null
+                    } }`,
                   },
         css: {
             preprocessorOptions: {
